updating the weeklysummary apis

+++ src/main/java/com/qentelli/employeetrackingsystem/repository/WeeklySummaryRepository.java
@@ -1,9 +1,9 @@
 package com.qentelli.employeetrackingsystem.repository;
-
-import com.qentelli.employeetrackingsystem.entity.WeeklySummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.qentelli.employeetrackingsystem.entity.WeeklySummary;
+
 @Repository
-public interface WeeklySummaryRepository extends JpaRepository<WeeklySummary, Long> {
+public interface WeeklySummaryRepository extends JpaRepository<WeeklySummary, Integer> {
 }
